+++ Sources/CodexBarCore/Vendored/CostUsage/CostUsageScanner.swift
@@ -87,6 +87,8 @@ enum CostUsageScanner {
             return CostUsageDailyReport(data: [], summary: nil)
         case .augment:
             return CostUsageDailyReport(data: [], summary: nil)
+        case .amp:
+            return CostUsageDailyReport(data: [], summary: nil)
         }
     }