+++ Sources/CodexBarWidget/CodexBarWidgetViews.swift
@@ -272,6 +272,7 @@ private struct ProviderSwitchChip: View {
         case .augment: "Augment"
         case .kimi: "Kimi"
         case .kimik2: "Kimi K2"
+        case .amp: "Amp"
         }
     }
 }
@@ -596,6 +597,8 @@ enum WidgetColors {
             Color(red: 254 / 255, green: 96 / 255, blue: 60 / 255) // Kimi orange
         case .kimik2:
             Color(red: 76 / 255, green: 0 / 255, blue: 255 / 255) // Kimi K2 purple
+        case .amp:
+            Color(red: 220 / 255, green: 38 / 255, blue: 38 / 255) // Amp red
         }
     }
 }